fix: error in tryInterpretAsLiteralElement
