+++ packages/dbml-core/src/parse/dbml/src/lib/parser/parser.ts
@@ -1,4 +1,8 @@
-import { extractIdentifierFromNode, isIdentifierNode } from '../utils';
+import {
+  extractIdentifierFromNode,
+  isIdentifierNode,
+  tryInterpretAsLiteralElement,
+} from '../utils';
 import { CompileError, CompileErrorCode } from '../errors';
 import { SyntaxToken, SyntaxTokenKind, isOpToken } from '../lexer/tokens';
 import Report from '../report';
@@ -286,56 +290,9 @@ export default class Parser {
       previousToken = this.previous();
     }
 
-    const maybeLiteralElement = this.tryInterpretAsLiteralElement(callee, args);
+    const maybeLiteralElement = tryInterpretAsLiteralElement(callee, args);
 
     return maybeLiteralElement || new FunctionApplicationNode({ callee, args });
-  }
-
-  private tryInterpretAsLiteralElement(
-    callee: ExpressionNode,
-    args: NormalFormExpressionNode[],
-  ): ElementDeclarationNode | undefined {
-    if (!isIdentifierNode(callee) || args.length === 0) {
-      return undefined;
-    }
-
-    const type = extractIdentifierFromNode(callee)!;
-
-    let attributeList: ListExpressionNode | undefined;
-    let alias: NormalFormExpressionNode | undefined;
-    let as: SyntaxToken | undefined;
-    let name: NormalFormExpressionNode | undefined;
-
-    const cpArgs = [...args];
-
-    const body = cpArgs.pop();
-    if (!(body instanceof BlockExpressionNode)) {
-      return undefined;
-    }
-
-    if (cpArgs[cpArgs.length - 1] instanceof ListExpressionNode) {
-      attributeList = cpArgs.pop() as ListExpressionNode;
-    }
-
-    const maybeAlias = cpArgs.pop();
-    const maybeAs = cpArgs[cpArgs.length - 1];
-    if (extractIdentifierFromNode(maybeAs)?.value === 'as') {
-      alias = maybeAlias;
-      as = extractIdentifierFromNode(cpArgs.pop());
-    } else {
-      name = maybeAlias;
-    }
-
-    return cpArgs.length === 0 ?
-      new ElementDeclarationNode({
-          type,
-          name,
-          as,
-          alias,
-          attributeList,
-          body,
-        }) :
-      undefined;
   }
 
   private normalFormExpression(): NormalFormExpressionNode {
